# api/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Asset Wallet Platform API
  version: "1.0"
  description: >
    Internal API for the digital asset wallet platform. Every mutating route requires
    bearer authentication and an Idempotency-Key header (AD-14). Errors use RFC 9457
    application/problem+json.
servers:
  - url: http://localhost:8080/v1
    description: Local dev (docker compose / make run)
  - url: /v1
security:
  - bearerAuth: []
paths:
  /customers:
    post:
      operationId: createCustomer
      summary: Create a customer and provision its per-asset accounts
      description: >
        Takes no request body — a customer has no caller-supplied fields in v1. Any body
        sent is ignored. The response is the created customer with its generated id.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          description: Bad request (e.g. missing Idempotency-Key header)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Idempotency-Key reused with a different request body
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /transfers:
    post:
      operationId: createTransfer
      summary: Move balance from one customer to another, ledger-only, no chain interaction
      description: >
        A single balanced journal entry (debit source, credit destination) commits
        atomically within one (chain, asset) pair — the same granularity as the
        provisioned per-customer accounts (FR4, AD-3, AD-4). No on-chain transaction is
        ever created.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      responses:
        "201":
          description: Transfer completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
        "400":
          description: >
            Bad request — missing Idempotency-Key, malformed body, invalid chain/asset,
            non-positive amount, or source equals destination
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Source or destination customer has no account for the requested (chain, asset)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: >
            Idempotency-Key reused with a different request body, or a concurrent
            request already used this key to write a journal entry
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: Source account balance is less than the requested amount
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /customers/{id}:
    get:
      operationId: getCustomer
      summary: Get a customer's own attributes, including its deposit address
      description: >
        Includes the customer's deposit address as an attribute of the customer
        resource (FR7) — never a route parameter of its own.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No customer with the given id exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /customers/{id}/balances:
    get:
      operationId: getCustomerBalances
      summary: Query a customer's current balance for each supported asset and chain
      description: >
        Each balance is derived from the ledger's postings, never a stored value (AD-3).
        Amounts are integer base units — wei for ETH, 6-decimal units for USDC — encoded
        as JSON strings, never numbers, to avoid float64 precision loss.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Customer balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalancesResponse"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No customer with the given id exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /customers/{id}/transactions:
    get:
      operationId: listCustomerTransactions
      summary: List a customer's transaction history, newest first
      description: >
        Reads generically from the cause-tagged journal (journal_entries joined to
        postings, restricted to this customer's own accounts) — no cause_type filter
        anywhere, so future cause types (deposit, withdrawal) appear automatically with
        no endpoint changes. amount is signed from this customer's own perspective:
        negative when debited, positive when credited.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          required: false
          description: Opaque page marker returned as nextCursor by a previous call. Omit for the first page.
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: Maximum number of transactions to return. Defaults to 20; values above 100 are clamped to 100.
          schema:
            type: integer
      responses:
        "200":
          description: One page of the customer's transaction history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionsResponse"
        "400":
          description: cursor does not decode to a valid page marker, or pageSize is present but not a positive integer
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No customer with the given id exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /customers/{id}/deposits:
    get:
      operationId: getCustomerDeposits
      summary: List a customer's pending deposits
      description: >
        Returns deposits in the observed or safe tier, both surfaced as status "pending"
        (Story 2.2 introduces the finalized/credited tiers and a non-pending status).
        Attribution is resolved via the customer's own deposit address (AD-8) — a
        deposit never carries a customer id of its own.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The customer's pending deposits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositsResponse"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No customer with the given id exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /withdrawals:
    post:
      operationId: createWithdrawal
      summary: Request a withdrawal, placing an immediate ledger-only hold on the requested amount
      description: >
        Reserves the requested amount the instant it is accepted (Story 3.2): a single
        balanced journal entry debits the customer's available account and credits its
        hold account for the same (chain, asset) pair — a plain ledger reclassification,
        exactly like an internal transfer, except both accounts belong to the same
        customer. No money leaves the customer and no on-chain transaction is created by
        this endpoint; signing and broadcasting are later stories (3.3-3.5). The
        destination address is validated for structural well-formedness only — never
        checked against a denylist or any other pre-signing policy, which is Story 3.3's
        job.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "201":
          description: Withdrawal accepted; its hold has been placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: >
            Bad request — missing Idempotency-Key, malformed body, invalid chain/asset,
            non-positive amount, or a destination address that is not a well-formed
            0x-prefixed 40-hex-character address
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No customer with the given id exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: >
            Idempotency-Key reused with a different request body, or a concurrent
            request already used this key to write a journal entry
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: >
            Available account balance is less than the requested amount, or (Story 3.3)
            covers the amount but not the amount plus the estimated withdrawal fee
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /withdrawals/{id}/approve:
    post:
      operationId: approveWithdrawal
      summary: Approve an awaiting-approval withdrawal (Story 3.3, FR17, NFR11)
      description: >
        Transitions a withdrawal from "awaiting-approval" to "approved", logging the
        approving operator's actor/timestamp/reason directly on the withdrawal row
        (NFR11). This system has no separate operator-identity/auth tier, so actor is
        caller-supplied in the request body, never derived from the request's auth
        context (a pre-existing, already-logged limitation of the shared-bearer-token
        model). Rejecting an already-approved or non-existent withdrawal is loud
        (404/409), never a silent no-op.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveWithdrawalRequest"
      responses:
        "200":
          description: Withdrawal approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: >
            Bad request — missing Idempotency-Key, malformed body, or missing/empty
            actor or reason
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No withdrawal with the given id exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: >
            The withdrawal is not currently awaiting approval (already approved, still
            created, or a concurrent request already won the approval race), or
            Idempotency-Key reused with a different request body
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Unexpected server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  schemas:
    Customer:
      type: object
      required:
        - id
        - createdAt
        - depositAddress
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        depositAddress:
          type: string
          pattern: '^0x[0-9a-fA-F]{40}$'
          minLength: 42
          maxLength: 42
          description: >
            The customer's counterfactual CREATE2 deposit address (FR6, FR7, AD-8) —
            identical on every supported EVM chain. Computed once at customer creation
            and persisted; never re-derived on request.
    Balance:
      type: object
      required:
        - chain
        - asset
        - balance
      properties:
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        balance:
          type: string
          description: >
            Integer base units (wei for ETH, 6-decimal units for USDC), encoded as a
            string — never a JSON number — to avoid float64 precision loss.
    BalancesResponse:
      type: object
      required:
        - balances
      properties:
        balances:
          type: array
          items:
            $ref: "#/components/schemas/Balance"
    TransferRequest:
      type: object
      required:
        - sourceCustomerId
        - destinationCustomerId
        - chain
        - asset
        - amount
      properties:
        sourceCustomerId:
          type: string
          format: uuid
        destinationCustomerId:
          type: string
          format: uuid
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        amount:
          type: string
          description: >
            Integer base units (wei for ETH, 6-decimal units for USDC), encoded as a
            string — never a JSON number — matching Balance's convention. Must be a
            positive integer.
    Transfer:
      type: object
      required:
        - id
        - sourceCustomerId
        - destinationCustomerId
        - chain
        - asset
        - amount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        sourceCustomerId:
          type: string
          format: uuid
        destinationCustomerId:
          type: string
          format: uuid
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        amount:
          type: string
        createdAt:
          type: string
          format: date-time
    WithdrawalRequest:
      type: object
      required:
        - customerId
        - chain
        - asset
        - amount
        - destinationAddress
      properties:
        customerId:
          type: string
          format: uuid
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        amount:
          type: string
          description: >
            Integer base units (wei for ETH, 6-decimal units for USDC), encoded as a
            string — never a JSON number — matching Balance's convention. Must be a
            positive integer.
        destinationAddress:
          type: string
          pattern: '^0x[0-9a-fA-F]{40}$'
          minLength: 42
          maxLength: 42
          description: >
            The on-chain address the withdrawal will eventually be sent to. Validated
            for structural well-formedness only (Story 3.2) — never checked against a
            denylist or any other pre-signing policy, which is Story 3.3's job.
    Withdrawal:
      type: object
      required:
        - id
        - customerId
        - chain
        - asset
        - amount
        - destinationAddress
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        amount:
          type: string
        destinationAddress:
          type: string
          pattern: '^0x[0-9a-fA-F]{40}$'
          minLength: 42
          maxLength: 42
        status:
          type: string
          enum: [created, awaiting-approval, approved]
          description: >
            "created" is a withdrawal's initial status, immediately superseded in the
            same request by Story 3.3's policy-check-and-route step — no withdrawal is
            ever observable at rest in this status via this API. "awaiting-approval" is
            reached when the requested amount exceeds its (chain, asset)'s configured
            approval threshold (FR17); "approved" is reached either automatically
            (amount at or below threshold) or via POST /withdrawals/{id}/approve. Later
            stories (3.4-3.5) add further status values as they add the transitions that
            produce them.
        approvedAt:
          type: string
          format: date-time
          description: >
            When this withdrawal was approved (Story 3.3, NFR11) — absent for a
            withdrawal still awaiting approval.
        approvedBy:
          type: string
          description: >
            The operator (or, for an auto-approved withdrawal, absent) who approved this
            withdrawal (Story 3.3, NFR11) — caller-supplied at approval time, since this
            system has no separate operator-identity/auth tier.
        approvalReason:
          type: string
          description: >
            The reason given for approving this withdrawal (Story 3.3, NFR11) — absent
            for an auto-approved withdrawal or one still awaiting approval.
        createdAt:
          type: string
          format: date-time
    ApproveWithdrawalRequest:
      type: object
      required:
        - actor
        - reason
      properties:
        actor:
          type: string
          description: >
            The operator approving this withdrawal (NFR11) — required and non-empty;
            this system has no separate operator-identity/auth tier to derive this from.
        reason:
          type: string
          description: The reason for approving this withdrawal (NFR11) — required and non-empty.
    Transaction:
      type: object
      required:
        - id
        - type
        - amount
        - chain
        - asset
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          description: >
            The journal entry's cause type, verbatim (e.g. "internal_transfer").
        amount:
          type: string
          description: >
            Integer base units (wei for ETH, 6-decimal units for USDC), encoded as a
            string — never a JSON number. Signed from this customer's own perspective:
            negative when debited, positive when credited.
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        status:
          type: string
          description: >
            "completed" for most cause types (e.g. "internal_transfer"); "credited" for a
            "deposit_credit" cause type (Story 2.2), the same terminal state as the
            underlying deposit.
        createdAt:
          type: string
          format: date-time
    TransactionsResponse:
      type: object
      required:
        - transactions
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        nextCursor:
          type: string
          description: Opaque page marker for the next page. Absent when this is the last page.
    Deposit:
      type: object
      required:
        - id
        - chain
        - asset
        - amount
        - txHash
        - status
        - tier
        - observedAt
      properties:
        id:
          type: string
          format: uuid
        chain:
          type: string
          enum: [base, arbitrum]
        asset:
          type: string
          enum: [eth, usdc]
        amount:
          type: string
          description: >
            Integer base units (wei for ETH, 6-decimal units for USDC), encoded as a
            string — never a JSON number — matching Balance's convention.
        txHash:
          type: string
          description: The on-chain transaction hash the deposit was observed in.
        status:
          type: string
          enum: [pending, orphaned]
          description: >
            "pending" for both the observed and safe tiers; "orphaned" once a reorg has
            evicted the deposit's block (Story 2.4) — Story 2.2's finalized/credited
            tiers are never surfaced on this endpoint (see tier).
        tier:
          type: string
          enum: [observed, safe, orphaned]
          description: This deposit's current position in the observed->safe->finalized->credited state machine (AD-6).
        observedAt:
          type: string
          format: date-time
    DepositsResponse:
      type: object
      required:
        - deposits
      properties:
        deposits:
          type: array
          items:
            $ref: "#/components/schemas/Deposit"
...
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string